+++ modules/tinymce/src/core/main/ts/api/OptionTypes.ts
@@ -169,6 +169,7 @@ interface BaseEditorOptions {
   placeholder?: string;
   preserve_cdata?: boolean;
   preview_styles?: false | string;
+  promotion?: boolean;
   protect?: RegExp[];
   readonly?: boolean;
   referrer_policy?: ReferrerPolicy;
@@ -300,6 +301,7 @@ export interface EditorOptions extends NormalizedEditorOptions {
   object_resizing: string;
   paste_as_text: boolean;
   preview_styles: string;
+  promotion: boolean;
   readonly: boolean;
   removed_menuitems: string;
   toolbar: boolean | string | string[] | Array<ToolbarGroup>;